Add confirm dialog for user to confirm destructive action as well as logic to conditionally delete the item based on their response

+++ src/components/List.jsx
@@ -68,8 +68,15 @@ export default function List() {
     };
   }, [userToken]);
 
-  const deleteHandler = async (item) => {
-    await deleteDoc(doc(db, userToken, item.id));
+  const deleteHandler = (item) => {
+    const deletionConfirmation = window.confirm(
+      `Are you sure you'd like to delete ${
+        item.data().item
+      } from your shopping list?`,
+    );
+    if (deletionConfirmation) {
+      deleteDoc(doc(db, userToken, item.id));
+    }
   };
 
   return (
